+++ server/db/index.ts
@@ -1,4 +1,4 @@
-const mongoose = require("mongoose");
+import mongoose from "mongoose";
 
 const userSchema = new mongoose.Schema({
     username: String,
@@ -12,10 +12,5 @@ const todoSchema = new mongoose.Schema({
     userId: String,
 });
 
-const User = mongoose.model('User', userSchema);
-const Todo = mongoose.model('Todo', todoSchema);
-
-module.exports = {
-    User,
-    Todo
-}
+export const User = mongoose.model('User', userSchema);
+export const Todo = mongoose.model('Todo', todoSchema);
