--- server/index.js
@@ -1,20 +1,0 @@
-
-const express = require("express");
-const app = express();
-const mongoose = require("mongoose");
-const port = 3000;
-const authRoutes = require("./routes/auth");
-const todoRoutes = require("./routes/todo");
-const cors = require("cors");
-const dotenv = require("dotenv").config();
-
-app.use(cors());
-app.use(express.json());
-app.use("/auth", authRoutes);
-app.use("/todo", todoRoutes);
-
-app.listen(port, () => {
-    console.log(`Example app listening at http://localhost:${port}`)
-})
-
-mongoose.connect(process.env.MONGO_URI, { dbName: "todo" });
