+++ server/routes/todo.ts
@@ -1,12 +1,12 @@
-const express = require('express');
-const { authenticateJwt, SECRET } = require("../middleware/index");
-const { Todo } = require("../db");
+import express from "express";
+import {authenticateJwt, SECRET } from "../middleware/index";
+import { Todo } from "../db";
 const router = express.Router();
 
 router.post('/todos', authenticateJwt, (req, res) => {
   const { title, description } = req.body;
   const done = false;
-  const userId = req.userId;
+  const userId = req.headers["userId"];
 
   const newTodo = new Todo({ title, description, done, userId });
 
@@ -21,7 +21,7 @@ router.post('/todos', authenticateJwt, (req, res) => {
 
 
 router.get('/todos', authenticateJwt, (req, res) => {
-  const userId = req.userId;
+  const userId = req.headers["userId"];
 
   Todo.find({ userId })
     .then((todos) => {
@@ -34,7 +34,7 @@ router.get('/todos', authenticateJwt, (req, res) => {
 
 router.patch('/todos/:todoId/done', authenticateJwt, (req, res) => {
   const { todoId } = req.params;
-  const userId = req.userId;
+  const userId = req.headers["userId"];
 
   Todo.findOneAndUpdate({ _id: todoId, userId }, { done: true }, { new: true })
     .then((updatedTodo) => {
@@ -48,4 +48,4 @@ router.patch('/todos/:todoId/done', authenticateJwt, (req, res) => {
     });
 });
 
-module.exports = router;
+export default router
