+++ server/tsconfig.json
@@ -1,0 +1,12 @@
+{
+  "compilerOptions": {
+    
+    "target": "es2016",
+    "module": "commonjs",
+    "esModuleInterop": true,                           
+    "forceConsistentCasingInFileNames": true,
+    "strict": true,
+    "skipLibCheck": true,
+    "outDir": "dist",
+  }
+}
